+++ codex-rs/tui/src/lib.rs
@@ -64,6 +64,7 @@ mod slash_command;
 mod status;
 mod status_indicator_widget;
 mod streaming;
+mod style;
 mod terminal_palette;
 mod text_formatting;
 mod tui;
